Introduce configuration file
This patch introduces a configuration file for Emissions API.
It will look for configuration files in the following order:

- ./emissionsapi.yml
- ~/emissionsapi.yml
- /etc/emissionsapi.yml

A configuration file template can be found at `etc/emissionsapi.yml`. To get
started, just copy this to the main project directory and adjust the values if
the defaults do not work for you.

This fixes #14
This is related to #11