Configure Root Logger for logging
Instead of creating a new logger for all our modules, we can simply change the
configuration of the root logger instead.

This does have the nice affect that every imported module also use this
configuration and therefore have the same formatting and log level.

Downside is that if the module is not executed via its generated binary but as
a module (python -m emissionsapi.*), the logged name will be __main__. If the
module is executed via its generated binary instead, the name matches the
module, e.g. emissionsapi.download.

Signed-off-by: Sven Haardiek <sven@haardiek.de>