Add pagination to the API
This patch introduces the query parameter `offset` and `limit` to introduce
pagination on our API.

This does not apply to the [JSON:API
Pagination](https://jsonapi.org/examples/#pagination), since both connexion and
the swagger-ui-bundle does not handle query parameters in the form
`page[limit]` well as explained in #89.

Nonetheless we should be able to extend this pagination later with links in the
response, if we wish to do that.

Closes #89