Introduce Point Query Parameter
This patch introduces points as query parameter. So it is possible to define a
point and the result is filtered for a small area around this point. This is
the implementation of the [NOZ user
story](https://github.com/emissions-api/project-notes/blob/master/user-stories/user-story-noz.md)

To make it more readable it also splits the database `get_*` function and the filter functions.
So we do not have to propagade the parameter through all those functions.