Requirements Rework
This patch splits the required dependencies in the `setup.py` and in the `requirements.txt`.

`setup.py` now contains a minimal set, which is needed to install emissions-api as a module.

`requirements.txt` contains a full list of dependencies including the
dependencies of dependencies and packages only needed for the installation like
`gunicorn`.