Remove download funtionality
Contrary to our initial assumption the download functionality of the
emissionsapi package is not used during production.

Processing and serving the files is mostly done separate from downloading the
data.
So we are using `sentinel5dl` to download and later process and serve the
result with `emissionsapi`.

Because of that this patch removes the whole download functionality from
emissionsapi.