Faster nc file listing
If processing a directory with a lot of files and most of them are already
processed, it is a lot faster to get all files from the database at once and
compare them in memory than to get every file separately from the database.
This consumes of course more memory, but IMHO is this okay for approx. 5000
files per year.

Co-Authored-By: Lars Kiesow <lkiesow@uos.de>