Support multiple products
This patch adds support for preprocessing and storing multiple products. It also
changes the API endpoints to be able to get data for different type of
products.

Since this also changes the REST API, it increases the version to version 2.

Also there is now an endpoint in the API to query for the available products.

Closes #154