Autoupdater
This patch adds a new functionality to emissiona-api. The autoupdater.

It is detecting the missing data from a defined starting point until the
present and will download and add it to the database. It is working in a
temporary directory and do not need a lot of disk space, since it is deleting
the data after processing it.

By default it does not check for new files in the already existing data range,
but this can be change by a command line parameter to force a comparision
between the already processed files and the files present by the ESA.