Always Honor `requirements.txt`
This patch removes the separately tracked list of dependencies from the
`setup.py`, instead loading the dependencies from the `requirements.txt`
so that we have a single point of truth for what dependencies we need.

This also means that the integration tests now actually honor the
specified dependencies.

This fixes #465