More Generic Metrics Backend
This patch basically provides an alternative to pull request #464.

This uses a more generic database table structure which will allow us to
use the same table for other metrics as well and not only for the
request counters.

We can use this to add more metrics and then just need to add an
additional collector exporting these for the metrics endpoint.

This patch does not modify the current behavior of the metrics endpoint.
It only changes the storage.